tools/catalog/requirements.txt
# Catalog tooling dependencies
# Install: pip3 install -r tools/catalog/requirements.txt
# Requires Python 3.11+

PyNaCl>=1.5.0        # ED25519 signing/verification (sign_indexes.py, verify_signatures.py)
requests>=2.31.0     # HTTP fetches (future tooling use)
